Cleaned changes and styled
