+++ client/src/components/Search/ServicesOutput/LargeService/LargeService.styles.ts
