+++ client/src/components/Search/ServicesOutput/SmallService/SmallService.styles.ts
