+++ client/src/pages/Demos/Demo.logic.ts
@@ -3,7 +3,6 @@ import { useContext, useEffect, useState } from "react";
 import { getSearchResults } from "../../helper/API";
 import { GET_DEMO_CONTENT } from "../../helper/CMS";
 import LanguageContext from "../../helper/LanguageContext";
-import { Service } from "../../types";
 
 const DemoLogic = () => {
     const [similarResults, setSimilarResults] = useState<any>({
@@ -13,26 +12,6 @@ const DemoLogic = () => {
             lng: undefined,
         },
     });
-
-    const getName = (service: Service) => {
-        let name =
-            language === "fr-CA" && service.nom !== ""
-                ? service.nom
-                : service.name;
-
-        name = name.replace(/[\u{0080}-\u{FFFF}]/gu, "");
-        return name;
-    };
-
-    const getDescription = (service: Service) => {
-        let desc =
-            language === "fr-CA" && service.description_fr !== ""
-                ? service.description_fr
-                : service.description;
-
-        desc = desc.replace(/[\u{0080}-\u{FFFF}]/gu, "");
-        return desc;
-    };
 
     const handleSimilar = (description: string) => {
         const searchBody = {
@@ -66,13 +45,12 @@ const DemoLogic = () => {
     const demoContent = data ? data.demos[0] : {};
 
     return {
+        language,
         similarResults,
         handleSimilar,
         useHandleDemoChange,
         error,
         demoContent,
-        getDescription,
-        getName,
     };
 };
